requirements.txt
tqdm
numpy
matplotlib
pandas
scikit_learn
swifter
wandb
wget
pytest
plotly
mapply
torchinfo
torchmetrics
einops
black>=22.6
coverage>=6
darglint>=1.8
isort>=5.10
pipreqs==0.4.11
pre-commit>=2.20
pytest-cov>=3
transformers
segmentation-models-pytorch
seaborn
pyod
jupyter
fitter
SciencePlots
scikit-image
opencv-python
latex
fastdup>=1.78 # 1.65
torch-lr-finder
faiss-gpu
cleanlab
lxml
imagehash
html5lib
openpyxl
pillow==9.0.0
